Toggle switch -Added a Toggle switch -Refactored file uploader to a seperate file -Created the url file component
