+++ audio-peeler-web-app/src/App.css
@@ -139,6 +139,8 @@
   box-shadow: 0px 5px 4px black;
 }
 
+
+
 #start-button:hover {
   box-shadow: 0px 7px 7px black;
   cursor: pointer;
