+++ audio-peeler-web-app/src/App.js
@@ -1,10 +1,10 @@
-import React from 'react';
+import React, {useMemo, useState} from 'react';
 import {
     BrowserRouter as Router,
     Switch,
     Route,
     Link
-  } from "react-router-dom";
+} from "react-router-dom";
 import './App.css';
 import './home/Home.css';
 import './mashup/Mashup.css';
@@ -12,35 +12,43 @@ import './samplers/Samplers.css';
 import Home from "./home/Home";
 import Mashup from './mashup/Mashup';
 import Samplers from './samplers/Samplers';
+import FileUrlContext from "./FileUrlContext";
 
 export default function App() {
+    const [john, setJohn] = React.useState(false);
+    const value = useMemo(() => ({john, setJohn}), [john]);
+
     return (
-      <Router>
-          <div id="header">
-              <button className="nav-button">Try the App!</button>
-        <Link to="/mashup">
-            <div className="nav-button" id="mashup-button">Mashup</div>
-        </Link>
-        <Link to="/samplers">
-            <div className="nav-button" id="samplers-button">Samplers</div>
-        </Link>
-        <Link to="/">
-            <div className="nav-button" id="home-button">Home</div>
-        </Link>
-          {/* A <Switch> looks through its children <Route>s and
+
+        <Router>
+            <div id="header">
+                <button className="nav-button">Try the App!</button>
+                <Link to="/mashup">
+                    <div className="nav-button" id="mashup-button">Mashup</div>
+                </Link>
+                <Link to="/samplers">
+                    <div className="nav-button" id="samplers-button">Samplers</div>
+                </Link>
+                <Link to="/">
+                    <div className="nav-button" id="home-button">Home</div>
+                </Link>
+                {/* A <Switch> looks through its children <Route>s and
               renders the first one that matches the current URL. */}
-          <Switch>
-            <Route path="/mashup">
-              <Mashup />
-            </Route>
-            <Route path="/samplers">
-              <Samplers />
-            </Route>
-            <Route path="/">
-              <Home />
-            </Route>
-          </Switch>
-          </div>
-      </Router>
+                <FileUrlContext.Provider value={value}>
+                <Switch>
+                    <Route path="/mashup">
+                        <Mashup/>
+                    </Route>
+                    <Route path="/samplers">
+                        <Samplers/>
+                    </Route>
+                    <Route path="/">
+                            <Home/>
+                    </Route>
+                </Switch>
+            </FileUrlContext.Provider>
+            </div>
+        </Router>
+
     );
-  }
+}
