+++ audio-peeler-web-app/src/FileUrlContext.js
@@ -1,0 +1,5 @@
+import {createContext, useContext} from "react";
+
+export const FileUrlContext = createContext([[], () =>{}])
+
+export default FileUrlContext;
