+++ selenium/src/main/resources/config.properties
@@ -2,5 +2,7 @@ vivaldiPath =C:\\Users\\frank\\AppData\\Local\\Vivaldi\\Application.exe
 ublockPath =C:\\Users\\frank\\AppData\\Local\\Vivaldi\\Application\\selenium\\extension_1_34_0_0.crx
 chromeDriver =C:\\Users\\frank\\AppData\\Local\\Vivaldi\\Application\\selenium\\chromedriver.exe
 profile =C:\\Users\\frank\\AppData\\Local\\Google\\Chrome\\User Data
+productionURL = https://audiopeeler.herokuapp.com
+localhostURL = http://localhost:3001
 email = oldnotescribe@gmail.com
 password = ALittleTeaPot1!
